# swagger.yaml
openapi: 3.0.0
info:
  title: Pet Shelter API
  version: 1.0.0
  description: Документація API для курсової роботи
  contact:
    name: Backend Developer
servers:
  - url: http://localhost:5000
    description: Local server

components:
  schemas:
    Animal:
      type: object
      required:
        - name
        - species
      properties:
        id:
          type: integer
          description: Унікальний ID
        name:
          type: string
          description: Кличка
        species:
          type: string
          description: Вид (CAT, DOG)
      example:
        id: 1
        name: Murzik
        species: CAT
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /api/auth/register:
    post:
      summary: Реєстрація нового користувача
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                firstName:
                  type: string
                lastName:
                  type: string
      responses:
        201:
          description: Користувача успішно створено

  /api/auth/login:
    post:
      summary: Вхід у систему (Отримання токена)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  default: user@example.com
                password:
                  type: string
                  format: password
                  default: password123
      responses:
        200:
          description: Успішний вхід. Повертає JWT токен.

  /api/animals:
    get:
      summary: Отримати список всіх тварин
      tags: [Animals]
      responses:
        200:
          description: Успішно отримано список
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Animal"
    post:
      summary: Додати нову тварину
      tags: [Animals]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - species
              properties:
                name:
                  type: string
                species:
                  type: string
                  enum: [CAT, DOG, OTHER]
                breed:
                  type: string
                age:
                  type: integer
      responses:
        201:
          description: Тварину успішно створено
  /api/applications:
    post:
      summary: Подати заявку на усиновлення (Спрацьовує черга)
      tags: [Applications]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - animalId
                - message
              properties:
                animalId:
                  type: integer
                  example: 1
                message:
                  type: string
                  example: "Я хочу забрати цього котика, бо він схожий на мене."
      responses:
        201:
          description: Заявку прийнято

  /api/applications/my:
    get:
      summary: Мої заявки
      tags: [Applications]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Список заявок